peer review - Yunke Nie
Great job!

The project's functionality and design are impressive. I loved the color design and the typography you chose. The information display is well-designed for me, very straightforward to display, and keep the whole page simple. The functionality part is perfect for me. I would say the design for those cards that showed on the page for users to confirm the appointment could improve a little bit: to design the "Go back" card different from the "Confirm" card. Also, maybe because of the deep color design, the detailed information there is not very noticeable for me.